Node: Prevent HTML entities from being double-escaped
This patch makes sure that any HTML entities contained within a string passed
to `Node.text` won’t be double-escaped, by making the string escaping code
read ahead when encountering an ampersand, and then filling in any pending
text following it once it has determined that it wasn’t the start of an
HTML entity. With this approach, the performance impact on the escaping code
is very low, since it retains close to `O(N)` time complexity.