Add `<picture>` and `<source>` element
This adds support for the picture html element. This enables support for
differnet images in night mode and dark mode version of a website.

There was on issue where a class `HTMLMediaContext` protocol was
already being used for Video/Audio. I renamed that existing class to
`HTMLMultiMediaContext` to avoid a name collision. This also keeps the
`<meida>` attribute in line with the properties in the HTML WC3 picture
spec[1].

https://www.w3schools.com/html/html_images_picture.asp