Prevent empty attribute values from being appended
When adding an attribute that should be appended to any existing one, first check if either
the existing or new attribute contain an empty value, and if so, don't perform the append.
Otherwise, we'll end up with extra whitespace within attribute values. Also, slight cleanup
of the attribute handling logic within `ElementRenderingBuffer`, to make the conditionals
a bit more clear.