Make Node.div API based on a protocol
Base the `Node.div` API on a new `HTMLDividableContext` protocol,
which enables that API to be shared between `HTML.BodyContext`
and `HTML.DescriptionListContext`. While this does come with a minor
downside of now requiring an explicit context whenever a `div` node
is created as a stand-alone value (when using `Node.div` directly, without
defining any child nodes within it), in practice that should be very rare, as
`div` nodes are most likely placed within a known context (such as within
a `body` node, or another node that inherits `HTML.BodyContext` from its
parent).